+++ src/hooks/useData.tsx
@@ -1,10 +1,11 @@
-import { ReactNode, createContext, useContext, useEffect, useState } from 'react'
+import { ReactNode, createContext, useContext, useEffect, useMemo, useState } from 'react'
 import { DataFrame, Obj, arrayDataFrame } from '../helpers/dataframe/index.js'
 
 interface DataContextType {
   data: DataFrame,
   key: string,
   version: number,
+  maxRowNumber: number
 }
 
 function getDefaultDataContext(): DataContextType {
@@ -12,6 +13,7 @@ function getDefaultDataContext(): DataContextType {
     data: arrayDataFrame([]),
     key: 'default',
     version: 0,
+    maxRowNumber: 0,
   }
 }
 
@@ -19,6 +21,7 @@ export const DataContext = createContext<DataContextType>(getDefaultDataContext(
 
 interface DataProviderProps<M extends Obj, C extends Obj> {
   data: DataFrame<M, C>,
+  maxRowNumber?: number,
   children: ReactNode
 }
 
@@ -26,7 +29,7 @@ function getRandomKey(): string {
   return crypto.randomUUID()
 }
 
-export function DataProvider<M extends Obj, C extends Obj>({ children, data }: DataProviderProps<M, C>) {
+export function DataProvider<M extends Obj, C extends Obj>({ children, data, maxRowNumber: propMaxRowNumber }: DataProviderProps<M, C>) {
   // We want a string key to identify the data.
   const [key, setKey] = useState<string>(getRandomKey())
   const [previousData, setPreviousData] = useState<DataFrame<M, C>>(data)
@@ -48,11 +51,16 @@ export function DataProvider<M extends Obj, C extends Obj>({ children, data }: D
     setVersion(0)
   }
 
+  const maxRowNumber = useMemo(() => {
+    return propMaxRowNumber ?? data.numRows
+  }, [propMaxRowNumber, data.numRows])
+
   return (
     <DataContext.Provider value={{
       data,
       key,
       version,
+      maxRowNumber,
     }}>
       {children}
     </DataContext.Provider>